+++ dert/src/gov/nasa/arc/dert/state/LandmarkState.java
@@ -31,7 +31,7 @@ public abstract class LandmarkState extends MapElementState {
 	 * @param pinned
 	 * @param position
 	 */
-	public LandmarkState(long id, MapElementState.Type mapElementType, String prefix, double size, Color color,
+	public LandmarkState(int id, MapElementState.Type mapElementType, String prefix, double size, Color color,
 		boolean labelVisible, ReadOnlyVector3 position) {
 		super(id, mapElementType, prefix, size, color, labelVisible);
 		this.position = new Vector3(position);
